fix: Use float32 as default data type

+++ src/logic.h
@@ -3,7 +3,7 @@
 
 #include "ndarray.h"
 
-double NDArray_All(NDArray *a);
+float NDArray_All(NDArray *a);
 int NDArray_ArrayEqual(NDArray *a, NDArray *b);
 
 #endif //PHPSCI_NDARRAY_LOGIC_H
